+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+# patch
+
+* Output file is relative to config.paths.public and defaultes to `config.paths.public/modernizr.js`
+* Scanned files are taken from config.files
+
 # modernizr-brunch 0.0.1 (Oct 20, 2015)
 
 * Initial version to push to NPM